+++ package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "@fast-csv/format": "^4.3.5",
     "@google-cloud/storage": "^5.7.0",
+    "@googlemaps/google-maps-services-js": "^3.1.14",
     "axios": "^0.21.1",
     "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
