+++ src/consts/index.ts
@@ -26,6 +26,6 @@ export const GCS_CLIENT_EMAIL = process.env.GCS_CLIENT_EMAIL || '';
 export const GCS_PRIVATE_KEY = process.env.GCS_PRIVATE_KEY || '';
 export const GCS_BASE_URL = 'https://storage.googleapis.com';
 
-export const GOOGLE_MAPS_API_KEY = process.env.GOOGLE_MAPS_API_KEY || '';
+export const GMAPS_API_KEY = process.env.GMAPS_API_KEY || '';
 
 export * from './upload';
